+++ scripts/git/Сделайте бэкап.ps1
@@ -1,2 +1,0 @@
-Compress-Archive -Path "G:\AI\IIChatTools\*" `
-                 -DestinationPath "IIChatTools-v1.0.1-tested.zip"