+++ scripts/setup/Set-ExecutionPolicy -Scope Process -ExecutionPolicy Bypass -Force.ps1
@@ -1,2 +1,0 @@
-Set-ExecutionPolicy -Scope Process -ExecutionPolicy Bypass -Force
-& ".\create_structure_1_0.ps1"