+++ scripts/setup/Запуск downloaders.ps1
@@ -1,3 +1,0 @@
-Set-ExecutionPolicy -Scope Process -ExecutionPolicy Bypass
-cd G:\AI\IIChatTools
-.\downloader2.ps1